requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
jinja2==3.1.3
google-genai>=1.0.0
pydantic==2.6.1
python-dotenv==1.0.0
# Document parsing libraries
pypdf2==3.0.1
python-docx==0.8.11
pytesseract==0.3.10
pillow==10.0.0